backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
Django==4.0.6
django-filter==21.1
django_debug_toolbar==3.8.1
djangorestframework==3.13.0
drf-yasg==1.21.10
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.15.3
six==1.17.0
sorl-thumbnail==12.10.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
